Create split prime string.c
Problem statement

Amazon's operations team needs an algorithm that can break out a list of products for a given order. The products in the order are listed as a string and the order items are represented as prime numbers. Given a string consisting of digits[0-9], count the number of ways the given string can be split into unique prime numbers using the entire string.

Note

The inputString does not contain leading zeroes.

Each number split of the given number must be in the range 2 to 10^6 inclusive.

Since the answer can be large, return the answer modulo (10000000007)

Example

Input

inputString = 11373

Output

6

Explanation

This string can be split into prime numbers in 6 different ways [11, 3, 7, 3], [113, 7, 3], [11, 3, 73], [11, 37, 3], [113, 73] and [11, 373].

So the output is 6.

Input format
The input to the function/method consists of an argument:

inputString, a string representing the input string.

Output format
Return an integer representing the number of ways the given string can be split into unique primes using the entire string.

Code constraints
1 <= length of inputStrings <= 105